app/agents/meta_agents/implementation_agent.py
from asyncio import subprocess
import os
from typing import Any, Dict, List
from concurrent.futures import ThreadPoolExecutor
import asyncio

from app.models.agent.task import Task
from app.models.agent.agent import Agent
from app.models.agent.belief import Belief
from app.models.agent.desire import Desire
from app.models.agent.intention import Intention
from app.models.agent.action import Action
from app.models.agent.agent_role import AgentRole
from app.agents.meta_agents.meta_agents import MetaAgent
from app.tools.reasoning_engine import ReasoningEngine
from app.tools.ontology_reasoner import OntologyReasoner
from app.models.knowledge.knowledge_base import KnowledgeBase
from app.tools.llm_manager import LLMManager
from app.models.knowledge.ontology.ontology import Ontology
from pydantic import BaseModel, Field


class ImplementationAgent(MetaAgent):
    def __init__(self, agent_id: str, name: str, api_key: str, llm_service: Any, agent_communication_service: Any):
        super().__init__(
            agent_id=agent_id,
            name=name,
            api_key=api_key,
            llm_service=llm_service,
            agent_communication_service=agent_communication_service
        )
        self.agent_type = "implementation"
        self.knowledge_base = KnowledgeBase()
        self.reasoning_engine = ReasoningEngine(self.knowledge_base, api_key)
        self.llm_manager = LLMManager()
        self.ontology = Ontology()  # Initialize with your implementation ontology
        self.ontology_reasoner = OntologyReasoner(self.llm_manager, self.ontology)
        self.executor = ThreadPoolExecutor(max_workers=10)
        self.active_agents = {}

        self._init_beliefs()
        self._init_desires()
        self._init_goals()
        self._init_plans()

    def _init_beliefs(self):
        self.add_belief("Proper implementation is crucial for MAS success")
        self.add_belief("Continuous integration and testing improve code quality")

    def _init_desires(self):
        self.add_desire("Implement a fully functional MAS", priority=10)
        self.add_desire("Maintain high code quality and standards", priority=9)

    def _init_goals(self):
        self.add_goal("Implement domain-specific MAS", priority=9)
        self.add_goal("Ensure robust error handling", priority=8)
        self.add_goal("Implement comprehensive testing suite", priority=8)

    def _init_plans(self):
        self.create_plan(
            self.goals[0].id,
            [
                "Set up development environment",
                "Implement agent infrastructure",
                "Implement individual agents",
                "Implement communication protocols",
                "Integrate components",
                "Perform unit testing",
                "Conduct integration testing"
            ]
        )

    async def reason(self):
        print("Starting reasoning process for implementation")
        
        current_beliefs = [belief.to_dict() for belief in self.beliefs]
        updated_beliefs = await self.reasoning_engine.reason({"beliefs": current_beliefs})
        
        for belief in updated_beliefs.get("beliefs", []):
            self.add_belief(belief["content"])

        new_knowledge = await self.ontology_reasoner.infer_new_knowledge()
        
        for concept in new_knowledge.get("new_concepts", []):
            self.add_belief(f"New implementation concept: {concept['name']} - {concept['description']}")
        
        for relationship in new_knowledge.get("new_relationships", []):
            self.add_belief(f"New implementation relationship: {relationship['name']} between {relationship['domain']} and {relationship['range']}")

        implementation_query = "What are the best practices for implementing a MAS?"
        implementation_practices = await self.ontology_reasoner.answer_query(implementation_query)
        self.add_belief(f"Best implementation practices: {implementation_practices}")

        if any(belief.description == "New implementation requirement received" for belief in self.beliefs):
            self.add_goal("Update implementation with new requirement", priority=8)
        
        if any(belief.description == "Implementation bug detected" for belief in self.beliefs):
            self.add_goal("Fix implementation bug", priority=10)

        current_beliefs = [belief.to_dict() for belief in self.beliefs]
        new_desires = await self.reasoning_engine.generate_desires(current_beliefs)
        
        for desire in new_desires:
            self.add_desire(desire["description"], desire["priority"])

        print("Reasoning process for implementation completed")

    async def plan(self):
        print("Starting planning process for implementation")
        for goal in self.goals:
            if goal.description == "Update implementation with new requirement":
                self.create_plan(
                    goal.id,
                    [
                        "Analyze new requirement",
                        "Identify affected components",
                        "Implement changes",
                        "Update unit tests",
                        "Perform integration testing"
                    ]
                )
            elif goal.description == "Fix implementation bug":
                self.create_plan(
                    goal.id,
                    [
                        "Reproduce bug",
                        "Identify root cause",
                        "Implement fix",
                        "Update tests",
                        "Verify fix"
                    ]
                )
        
        current_state = self.get_current_state()
        optimized_plan = await self.reasoning_engine.reason_and_plan(self.goals[0].description, current_state)
        
        if optimized_plan.get("plan"):
            self.update_plan(self.goals[0].id, optimized_plan["plan"].steps)
        
        print("Planning process for implementation completed")

    async def execute(self):
        print("Starting execution process for implementation")
        for plan in self.plans:
            for task in plan.steps:
                if task.status == "pending":
                    print(f"Executing task: {task.description}")
                    try:
                        execution_result = await self.reasoning_engine.simulate_action(task.description, self.get_current_state())
                        await self.execute_task(task)
                        self.update_task_status(task.id, "completed")
                        
                        for key, value in execution_result.items():
                            self.add_belief(f"Implementation task result - {key}: {value}")
                    except Exception as e:
                        print(f"Error executing implementation task {task.description}: {str(e)}")
                        self.update_task_status(task.id, "failed")
        print("Execution process for implementation completed")

    def get_current_state(self) -> Dict[str, Any]:
        return {
            "beliefs": [belief.to_dict() for belief in self.beliefs],
            "desires": [desire.to_dict() for desire in self.desires],
            "goals": [goal.to_dict() for goal in self.goals],
            "plans": [plan.to_dict() for plan in self.plans],
        }

    async def run(self):
        while True:
            await self.reason()
            await self.plan()
            await self.execute()
            await asyncio.sleep(1)  # Adjust the sleep time as needed

    async def execute_task(self, task: Task):
        if task.description == "Set up development environment":
            await self.setup_development_environment()
        elif task.description == "Implement agent infrastructure":
            await self.implement_agent_infrastructure()
        elif task.description == "Implement individual agents":
            await self.implement_individual_agents()
        elif task.description == "Implement communication protocols":
            await self.implement_communication_protocols()
        elif task.description == "Integrate components":
            await self.integrate_components()
        elif task.description == "Perform unit testing":
            await self.perform_unit_testing()
        elif task.description == "Conduct integration testing":
            await self.conduct_integration_testing()
        else:
            print(f"Executing generic task: {task.description}")

    async def implement_mas(self, infrastructure_design: Dict[str, Any], agent_implementations: Dict[str, Any], communication_infrastructure_design: Dict[str, Any], communication_protocols: Dict[str, Any]) -> Dict[str, Any]:
        implementation_result = {
            "status": "in_progress",
            "components": {}
        }

        tasks = [
            self.executor.submit(self.implement_agent_infrastructure, infrastructure_design),
            self.executor.submit(self.implement_individual_agents, agent_implementations),
            self.executor.submit(self.implement_communication_infrastructure, communication_infrastructure_design),
            self.executor.submit(self.implement_communication_protocols, communication_protocols)
        ]

        for future in asyncio.as_completed(tasks):
            result = await future
            implementation_result["components"].update(result)

        implementation_result["status"] = "completed"
        return implementation_result

    async def setup_development_environment(self):
        print("Setting up development environment...")
        # Implement logic to set up the development environment
        # This could include installing dependencies, setting up virtual environments, etc.

    async def implement_agent_infrastructure(self, infrastructure_design: Dict[str, Any]) -> Dict[str, Any]:
        print("Implementing agent infrastructure...")
        try:
            # Implement the agent infrastructure based on the design
            # This is a placeholder implementation
            with open("src/agent_infrastructure.py", "w") as f:
                f.write(f"# Agent Infrastructure\n# Design: {infrastructure_design}\n\nclass AgentInfrastructure:\n    pass")
            return {"agent_infrastructure": "implemented"}
        except Exception as e:
            print(f"Error implementing agent infrastructure: {str(e)}")
            return {"status": "failed", "details": str(e)}

    async def implement_individual_agents(self, agent_implementations: Dict[str, Any]) -> Dict[str, Any]:
        print("Implementing individual agents...")
        implemented_agents = {}
        try:
            for agent_name, agent_spec in agent_implementations.items():
                # Implement each agent based on its specification
                # This is a placeholder implementation
                with open(f"src/agents/{agent_name}.py", "w") as f:
                    f.write(f"# Agent: {agent_name}\n# Spec: {agent_spec}\n\nclass {agent_name}:\n    pass")
                implemented_agents[agent_name] = "implemented"
            return implemented_agents
        except Exception as e:
            print(f"Error implementing individual agents: {str(e)}")
            return {"status": "failed", "details": str(e)}

    async def implement_communication_infrastructure(self, communication_infrastructure_design: Dict[str, Any]) -> Dict[str, Any]:
        print("Implementing communication infrastructure...")
        try:
            # Implement the communication infrastructure based on the design
            # This is a placeholder implementation
            with open("src/communication/infrastructure.py", "w") as f:
                f.write(f"# Communication Infrastructure\n# Design: {communication_infrastructure_design}\n\nclass CommunicationInfrastructure:\n    pass")
            return {"communication_infrastructure": "implemented"}
        except Exception as e:
            print(f"Error implementing communication infrastructure: {str(e)}")
            return {"status": "failed", "details": str(e)}

    async def implement_communication_protocols(self, communication_protocols: Dict[str, Any]) -> Dict[str, Any]:
        print("Implementing communication protocols...")
        try:
            # Implement the communication protocols
            # This is a placeholder implementation
            with open("src/communication/protocols.py", "w") as f:
                f.write(f"# Communication Protocols\n# Protocols: {communication_protocols}\n\nclass CommunicationProtocols:\n    pass")
            return {"communication_protocols": "implemented"}
        except Exception as e:
            print(f"Error implementing communication protocols: {str(e)}")
            return {"status": "failed", "details": str(e)}

    async def integrate_components(self, components: Dict[str, Any] = None) -> bool:
        print("Integrating components...")
        try:
            with open("src/main.py", "w") as f:
                f.write("""
import asyncio
from src.agents.base_agent import BaseAgent
from src.communication.message_bus import message_bus
from src.communication.protocols import MessageType, create_message

# Import individual agent classes here

async def run_agent(agent):
    while True:
        agent.run()
        await asyncio.sleep(0.1)

async def main():
    # Initialize agents
    agents = [
        # Instantiate your agents here
    ]

    # Start all agents
    await asyncio.gather(*[run_agent(agent) for agent in agents])

if __name__ == "__main__":
    asyncio.run(main())
""")
            print("Components integrated successfully.")
            return True
        except Exception as e:
            print(f"Error integrating components: {str(e)}")
            return False

    async def perform_unit_testing(self):
        print("Performing unit testing...")
        try:
            # Create a sample unit test
            with open("tests/test_base_agent.py", "w") as f:
                f.write("""
import unittest
from src.agents.base_agent import BaseAgent

class TestBaseAgent(unittest.TestCase):
    def test_base_agent_initialization(self):
        agent = BaseAgent("test_id", "Test Agent")
        self.assertEqual(agent.agent_id, "test_id")
        self.assertEqual(agent.name, "Test Agent")
        self.assertEqual(agent.beliefs, [])
        self.assertEqual(agent.desires, [])
        self.assertEqual(agent.intentions, [])

if __name__ == '__main__':
    unittest.main()
""")
            # Run the unit tests
            result = subprocess.run(["python", "-m", "unittest", "discover", "-v", "-s", "tests"], capture_output=True, text=True)
            print(result.stdout)
            if result.returncode != 0:
                print(f"Unit tests failed: {result.stderr}")
            else:
                print("Unit tests completed successfully.")
        except Exception as e:
            print(f"Error performing unit testing: {str(e)}")

    async def conduct_integration_testing(self):
        print("Conducting integration testing...")
        try:
            # Create a sample integration test
            with open("tests/test_integration.py", "w") as f:
                f.write("""
import unittest
import asyncio
from src.communication.message_bus import message_bus
from src.communication.protocols import MessageType, create_message

class TestIntegration(unittest.TestCase):
    def test_message_bus(self):
        async def test_publish_subscribe():
            received_messages = []
            
            async def subscriber(message):
                received_messages.append(message)
            
            message_bus.subscribe("test_topic", subscriber)
            await message_bus.publish("test_topic", "Test message")
            
            # Wait for the message to be processed
            await asyncio.sleep(0.1)
            
            self.assertEqual(len(received_messages), 1)
            self.assertEqual(received_messages[0], "Test message")
        
        asyncio.run(test_publish_subscribe())

if __name__ == '__main__':
    unittest.main()
""")
            # Run the integration tests
            result = subprocess.run(["python", "-m", "unittest", "discover", "-v", "-s", "tests"], capture_output=True, text=True)
            print(result.stdout)
            if result.returncode != 0:
                print(f"Integration tests failed: {result.stderr}")
            else:
                print("Integration tests completed successfully.")
        except Exception as e:
            print(f"Error conducting integration testing: {str(e)}")
